"Resolve segmentation issues"
